fix: remove authn/z proxy
- frontier started as an IAP that can do authentication and
authorization before any request can hit any of the backend
microservices. It supported http1.1 & 2, REST based as well as
grpc protocols but to ensure proxy is working end to end for all
corner cases it was hard to maintain.
- so it was decided to remove this design and only keep frontier
as a standalone service and use a dedicated proxy like kong/trafiek
for this usecase.

Signed-off-by: Kush Sharma <thekushsharma@gmail.com>