fix: list only explicitly added org members
- currently org to user relation ships are calculated dynamically
and all direct or indirect access controls will be evaluated but
after this change only users who are added as a member or given
access via a policy will be listed. For example, super users will
who have access to whole platform will not be part of the list if
they don't have explicit access to the org.

Signed-off-by: Kush Sharma <thekushsharma@gmail.com>