feat: pass stripe test clock id while creating customer
Stripe allows simulating test clocks to test subscriptions, payments
and invoices. This clock needs to be created from stripe dashboard
and clock id needs to be passed as a request header while creating a
new billing customer account for stripe to use this simulated clock.
Once the customer is created, there is no need to pass this header
 and all of its subsequent subscriptions will automatically be part
of the simulated clock.

Keep in mind once the clock expires, all of it's resources also gets
expired.

Header Name: X-Stripe-Test-Clock

Signed-off-by: Kush Sharma <thekushsharma@gmail.com>