fix: billing account with no state in db were failing entitlement checks (#647)
Billing account were not adding active state in db by default. We
started adding active state later causing use to make the change
backward compatible. This change while being made backward compatible
used a where condition in db which although looked like it is using
a `AND` between org and state filter, it was using an `OR`. The
underlying library `goqu` didn't make it obvious to see from its
methods.

Signed-off-by: Kush Sharma <thekushsharma@gmail.com>