fix: do not sync deleted test billing customer/subscriptions
If a customer is deleted from billing provider but not in stripe,
it was not getting marked deleted earlier, not it will be marked
disabled correctly.
Similarly if a subscription is deleted from billing provider
which was marked as a test resource will also be marked canceled
correctly in frontier and sync will no longer happen for these
resources.
Additionally, 0 amount billing usage reporting is disabled.

Signed-off-by: Kush Sharma <thekushsharma@gmail.com>