fix: while updating role, do full refresh of permissions
Currently while updating role, frontier used to identify the deleted
permissions from role and only remove them which are needed. One
problem with this is if in case someone manually modifies such roles
and the permissions gets out of sync, it will not try to correct them.
Now it deletes all existing permissions and create the new requested
on every update call.

Signed-off-by: Kush Sharma <thekushsharma@gmail.com>